refactor: extract register commands func
