+++ src/commands/index.ts
@@ -1,0 +1,5 @@
+export * from './baseCommand';
+
+export * from './searchFilename';
+export * from './searchRepository';
+export * from './searchFileContent';
